+++ src/components/bookTable/bookTable.js
@@ -7,8 +7,7 @@ import {
   useDeleteBookMutation,
 } from 'redux/books/booksApi';
 import { useTranslation } from 'react-i18next';
-import RatingBook from 'components/RatingBook';
-import NestingModal from 'components/RatingBook/RatingModal/NestingModal/NestingModal';
+import RatingBookWrapper from 'components/RatingBookWrapper';
 
 export default function BookTable() {
   const { data } = useGetAllBooksQuery();
@@ -18,12 +17,6 @@ export default function BookTable() {
   const status = e => {
     const status = data?.payload.books.some(book => book.status === e);
     return status;
-  };
-
-  const [showModal, setShowModal] = useState(false);
-
-  const toggleModal = () => {
-    setShowModal(!showModal);
   };
 
   return (
@@ -121,24 +114,11 @@ export default function BookTable() {
                               <use href={`${icons}#white_star`}></use>
                             </svg>
                           )}
-                          <button
-                            type="button"
+                          <RatingBookWrapper
                             id={_id}
-                            className={s.btn}
-                            onClick={toggleModal}
-                          >
-                            {t('resume')}
-                          </button>
-                          {showModal && (
-                            <NestingModal onClose={toggleModal}>
-                              <RatingBook
-                                toggleModal={toggleModal}
-                                id={_id}
-                                resume={resume}
-                                rating={rating}
-                              />
-                            </NestingModal>
-                          )}
+                            resume={resume}
+                            rating={rating}
+                          />
                         </td>
                       </tr>
                     )
